backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
moviepy==1.0.3
stability-sdk==0.8.4
elevenlabs==0.2.26
pillow==10.1.0
python-jose[cryptography]==3.3.0
pydantic==2.5.2
pydantic-settings==2.1.0
pytest==7.4.3
httpx==0.25.2
python-magic==0.4.27
nltk==3.8.1
grpcio==1.53.0
protobuf==4.21.12 